+++ src/lib/models.ts
@@ -72,6 +72,19 @@ const getAllPropertiesById = async (ids: any[]): Promise<ListingType[]> => {
 
   return data
 }
+const getFavourites = async (userId: any) => {
+  // const id = await getUserId()
+  const { data, error } = await supabaseClient
+    .from('favourites')
+    .select()
+    .eq('user_id', userId)
+  if (error) {
+    console.error('Error getting favourites data', error)
+    throw error
+  }
+  console.log('favourite data', data)
+  return data
+}
 
 const getAllProperties = async (): Promise<ListingType[]> => {
   const { data, error } = await supabaseClient.from('property').select(
@@ -178,4 +191,5 @@ export {
   getRoleByDescription,
   getAllRoles,
   getAllPropertiesById,
+  getFavourites,
 }
